Fix the cleanup of subprocesses after each AT
