+++ acceptance-tests/tests/src/test/java/org/hyperledger/besu/tests/acceptance/permissioning/AccountSmartContractPermissioningAcceptanceTestBase.java
@@ -30,6 +30,8 @@ import org.hyperledger.besu.tests.acceptance.dsl.transaction.perm.AccountSmartCo
 import java.io.IOException;
 import java.util.List;
 
+import org.junit.jupiter.api.AfterEach;
+
 class AccountSmartContractPermissioningAcceptanceTestBase extends AcceptanceTestBase {
 
   private final AccountSmartContractPermissioningTransactions smartContractAccountPermissioning;
@@ -77,6 +79,7 @@ class AccountSmartContractPermissioningAcceptanceTestBase extends AcceptanceTest
     }
   }
 
+  @AfterEach
   @Override
   public void tearDownAcceptanceTestBase() {
     permissionedCluster.stop();
